fix warning

+++ disorderfs.cpp
@@ -121,7 +121,7 @@ bool compare_ctime_dirents(Ctime_Dirent_pair a, Ctime_Dirent_pair b){
 
 // PRE: dirents.size() == sorted_interim.size()
 void overwrite_dirents(std::unique_ptr<Dirents>& dirents, std::vector<Ctime_Dirent_pair>& sorted_interim){
-    for(int i = 0; i < sorted_interim.size(); i++){
+    for(long unsigned int i = 0; i < sorted_interim.size(); i++){
         *(dirents->begin() + i) = (sorted_interim.begin() + i)->second;
     }
 };
